+ elaborated_module; updated ast
