+++ dbuf-core/src/lib.rs
@@ -1,1 +1,3 @@
-pub mod ast;
+pub mod elaborated_module;
+pub mod expression;
+pub mod parsed_module;
